+++ lib/public/StormByte/network/server.hxx
@@ -38,6 +38,10 @@ namespace StormByte::Network {
 		class Server;	///< Forward declaration
 	}
 
+	namespace Detail {
+		class Session;	///< Forward declaration
+	}
+
 	/**
 	 * @class Server
 	 * @brief Abstract application server.
@@ -116,6 +120,7 @@ namespace StormByte::Network {
 			Connection::HandlerType m_wakeup_read;												///< Wakeup read handle
 			Connection::HandlerType m_wakeup_write;												///< Wakeup write handle
 			std::unordered_map<std::string, std::shared_ptr<Connection::Client>> m_clients;		///< Active clients
+			std::unordered_map<std::string, std::shared_ptr<Detail::Session>> m_sessions;	///< Active parser sessions
 			std::unordered_map<std::string, std::thread> m_handle_msg_threads;						///< Per-client workers
 			std::mutex m_mutex;																		///< Protects client maps